Added Linux Source Files.
Used the MSVC Source files as a base and made changes that were needed to compile the project correctly.
There is no guarantee that this version of the API works as intended. Unit testing and an integration test are required.

Starting to move away from C Style file and string IO methods in favor of more standard C++ methods.